completado

+++ database/migrations/2025_11_25_141414_create_membresias_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('membresia', function (Blueprint $table) {
             $table->id();
             $table->string('nombre');     
-            $table->integer('precio');    
+            $table->float('precio');    
             $table->integer('duracion_dias');
             $table->timestamps();
         });
